requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
pydantic-settings
python-dotenv
pytest
pytest-mock 

msal
PyJWT
requests
openpyxl
